anticipated WETH claiming + move claiming logic to the Accumulator
ZeroLend will add WETH rewards in the near future, we already anticipated this new feature.
The Accumulator now calls the `execute` function on the locker to have more flexibility to claim fees and avoid having dead source code in the case where ZeroLend would upgrade their contract to a an incompatible version.