fix(ci): fix/improve CI workflows
This repository now relies on a dependency that enforces the use of Yarn
and attempts to run Yarn scripts during the installation step. To prevent
this, I created a .yarnrc.yml file at the root of the monorepo that
disables scripts for the @pendle/v2-sy package. However, we still need
to install Yarn on the CI runners for the workflows to function properly.
This commit uses the corepack feature to automatically install Yarn on
the CI runners.